Partial commit

+++ src/main/java/com/mimacom/calculator/controller/CalculatorSimpleRestController.java
@@ -9,6 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/calculator")
 public class CalculatorSimpleRestController {
@@ -18,22 +20,22 @@ public class CalculatorSimpleRestController {
     private CalculatorRestControllerSimpleService calculatorRestControllerSimpleService;
 
     @PostMapping(value = "/add/{valuea}/{valueb}", produces = APPLICATION_JSON)
-    public ResponseEntity<Integer> add(@PathVariable String valuea, @PathVariable String valueb) {
+    public ResponseEntity<BigDecimal> add(@PathVariable String valuea, @PathVariable String valueb) {
         return new ResponseEntity<>(calculatorRestControllerSimpleService.add(valuea, valueb), HttpStatus.OK);
     }
 
     @PostMapping(value = "/subtract/{valuea}/{valueb}" , produces = APPLICATION_JSON)
-    public ResponseEntity<Integer> subtract(@PathVariable String valuea, @PathVariable String valueb) {
+    public ResponseEntity<BigDecimal> subtract(@PathVariable String valuea, @PathVariable String valueb) {
         return new ResponseEntity<>(calculatorRestControllerSimpleService.subtract(valuea, valueb), HttpStatus.OK);
     }
 
     @PostMapping(value = "/multiply/{valuea}/{valueb}" , produces = APPLICATION_JSON)
-    public ResponseEntity<Integer> multiply(@PathVariable String valuea, @PathVariable String valueb) {
+    public ResponseEntity<BigDecimal> multiply(@PathVariable String valuea, @PathVariable String valueb) {
         return new ResponseEntity<>(calculatorRestControllerSimpleService.multiply(valuea, valueb), HttpStatus.OK);
     }
 
     @PostMapping(value = "/divide/{valuea}/{valueb}" , produces = APPLICATION_JSON)
-    public ResponseEntity<Double> divide(@PathVariable String valuea, @PathVariable String valueb) {
+    public ResponseEntity<BigDecimal> divide(@PathVariable String valuea, @PathVariable String valueb) {
         return new ResponseEntity<>(calculatorRestControllerSimpleService.divide(valuea, valueb), HttpStatus.OK);
     }
 }
